Data file Created, Object initialization templates
