--- src/utils/alignment.js
@@ -1,11 +1,0 @@
-export const alignments = {
-  LG: "Lawful Good",
-  LN: "Lawful Neutral",
-  LE: "Lawful Evil",
-  NG: "Neutral Good",
-  N: "True Neutral",
-  NE: "Neutral Evil",
-  CG: "Chaotic Good",
-  CN: "Chaotic Neutral",
-  CE: "Chaotic Evil",
-};
