--- src/utils/charAbilityScoreInitialState.js
@@ -1,14 +1,0 @@
-export const charAbilityScoreInitialState = {
-  str: 10,
-  dex: 10,
-  con: 10,
-  int: 10,
-  wis: 10,
-  cha: 10,
-  strMod: 0,
-  dexMod: 0,
-  conMod: 0,
-  intMod: 0,
-  wisMod: 0,
-  chaMod: 0,
-};
