--- src/utils/charHeaderInitialState.js
@@ -1,20 +1,0 @@
-export const charHeaderInitialState = {
-  name: "",
-  alignment: "LG",
-  player: "",
-  campaign: "",
-  class: "",
-  level: 1,
-  race: "",
-  deity: "",
-  homeland: "",
-  size: "",
-  gender: "",
-  age: "",
-  height: 0,
-  metricHeight: 0,
-  weight: 0,
-  metricWeight: 0,
-  hair: "",
-  eyes: "",
-};
