--- src/utils/drBypassTypes.js
@@ -1,1 +1,0 @@
-export const drBypassTypes = ["--", "Magic", "Cold Iron", "Bludgeoning", "Piercing", "Slashing", "Epic", "Adamantine", "Silver", "Good", "Evil", "Chaos", "Law"]
