--- src/utils/healthInitialState.js
@@ -1,10 +1,0 @@
-export const healthInitialState = {
-  hpCurrent: 0,
-  hpMax: 0,
-  wounds: 0,
-  woundsNonLethal: 0,
-  initiative: 0,
-  DR: 0,
-  SR: 0,
-  drBypass: "",
-}
